test(integration): consolidate 5 async tests with tokio seed
Merge async_alloc, async_main_return_type, async_nested_select,
async_self_time, and async_tokio into a single async_integration.rs
test binary. All 5 tests share a pre-compiled tokio seed via
prepopulate_deps, cutting redundant tokio compilation from 5x to 1x.
